Update package-lock.json to lockfileVersion 2
The new `package-lock.json` format is new in npm 7, which we use
everywhere, because it's a requirement to use npm workspaces. Since
every build will use npm 7, it will update this file for every build,
and that's not necessary and may have people think something is wrong
when it's not.