Link npm packages for nodejs test app
Previously the `package.json` pointed to the appsignal-nodejs workspace
for the Node.js integration packages. I'm not sure how it figured which
packages to use in that scenario, if not all packages in the project. I
changed this to specify both packages that are necessary: nodejs and
nodejs-ext.

This fixed the issue for me where the extension loading failed every
time, and thus the diagnose specs. By linking the packages to the
packages in the appsignal-nodejs workspace this issue is fixed for me.

This required the runners to have multiple setup commands, so I renamed
it to `setup_commands` from `setup_command`.

It now also checks if those setup commands fail or not and fail the test
if they do. I noticed the npm install script failed a couple time while
changing this and those failures went unnoticed.