Log full actual output on expect_output failure
This will help debug the assertion failure, as you can see the full
output that was asserted again. If one line fails, the assertion failure
message can be something like `\n` didn't match `<complex regex>`.
Without the context of all lines that were asserted this can be
difficult to debug.

Update `expect_output` helper to print the full actual output that was
asserted against (so far) on failure.