Always end the appsignal.log with an empty line
The log file should always end with an empty line. When it does not, the
logger will write the next line at the end of the last logged line.

```
line 1line 2
```

But because the logger will always end the log line with a line break,
our dummy log should too. The next time the logger will log a line it
will look like this instead:

```
line 1
line 2
```