Store install reports without indentation
The diagnose reports were stored with indentation as defined in the
diagnose spec file.

Use the HEREDOC syntax along with the tilde `~` to remove the
indentation of the report files. This way it can't be parsed
incorrectly. I didn't see that happen, but it was confusing why the
files were formatted like this and we should avoid it if we can.