Replace prepare step with callbacks
The prepare step can be split up into before and after setup callbacks.
This gives more control when exactly the setup is happening.

This is useful and possible necessary for the Node.js integration at
least, since the setup commands may overwrite the install report (as it
runs the installer itself). I'm in the process of potentially rewriting
the install script for Node.js to not skip writing an install report for
local installations, meaning the installer will always write a report.
If it does we need to overwrite the install report after the
installation again.

Since the before and after callbacks need to be run in very precise
steps along with the rest of the setup, remove the explicit `prepare`
call from the runners.