Fix the Node.js install report location
## Background

The Node.js integration's install report is now stored with a more
unique filename based on the app install path. This way it doesn't
overwrite the install report if multiple apps on the same host use the
AppSignal package.

https://github.com/appsignal/appsignal-nodejs/pull/412

## Problem

The diagnose tests wrote the install report to an outdated location and
thus wasn't actually used by the diagnose tests. This came to light when
changing the install script of the Node.js integration to always write
an install report, also for local installs. This change has not been
merged at time of writing, but highlighted this issue that is a problem
now. Previously it worked accidentally because the installer wrote a new
report upon installation and that happened to match the expected ouput.

## Solution

Adapt the new naming scheme as introduced in the previously linked
Node.js integration PR, but this time in Ruby.

Closes #7