Refactor diagnose test runner
I found it difficult to debug failure scenarios in the test suite. This
change hopes to make it a little better, but still the output is less
than ideal sometimes comparing an empty line against an expected regex.

## The refactor

Rather than using `readline` to read the diagnose tool output when we
need the next line, collect all the output beforehand. This makes it
easier to debug the full output when a failure occurs.

We now pass the full output with `Output#to_s` to the `expect` command
to print the full output on failure. We did something similar before,
but that only printed the _collected output so far_ in the test suite.

I split the Runner into a separate Output class and moved all the output
related behavior there. This contains the full output we can scan and
print.

# No need to "kill"

Because the test "wait"s for the diagnose tool to be ready beforehand,
we no longer need to "kill" the process at the end of the spec.

## Performance

In my limited testing I noticed no performance degradation. The diagnose
command itself doesn't take long to execute. In fact it ran a fraction
of a second faster (without `npm install` setup and such), from `2.9`
seconds to `2.6` seconds.

This was no professional benchmark, just a test to see if and how much
slower it would be and I see no evidence of that currently.