Only run test command in test directory
The `Dir.chdir` command changes the current working directory for the
entire test suite, not just when the runner is executed. This can create
some weird behavior for commands or RSpec expecting the directory to
remain the same.

Since we don't need to run the entire test suite in that test directory,
only run setup commands and the command run by the runner in this test
working directory.