Split runner into a separate file
Move the Runner class to its own file so that the diagnose_spec file is
a bit easier to read and find your way around.

I've updated all the paths in the new runner file to account for the
change in location, moving everything up one level to navigate out the
`support` sub directory.