Add runner helper for tests
Don't fetch and initialize the runner instance in every test itself, but
add a helper for it with additional error handling.

I removed the "ruby" default, because it burned me a couple times when I
hadn't set the LANGUAGE environment variable. It's better to just have
it error when that happens and have all integrations behave the same.