Test missing install report
When no agent and extension install report is present on the file
system, test if the diagnose command prints the error correctly.

To communicate to the runner that the install report should be created I
added an options argument to the Runner class, which can be expanded
with other future options.