Refactor install report behavior for runners
Move the writing and removing of the install report to the base runner
class so that individual runners don't have to implement it.

The install report path is unique, so that will remain in the runner for
the specific language.

I disabled the RuboCop rule for now. The refactor of the `Runner#run`
method is something we can do later on as well.