Use File.expand_path instead of File.join
Add a helper called `project_path` to provide the path out of the
`spec/support` directory and make creating paths to the app paths or the
integrations easier. We won't have to consider the location of the
support file every time we create a path to either of those.

Use `File.expand_path` to create an absolute path instead of a mix of an
absolute and relative path.

```
# Before
/appsignal-ruby/spec/diagnose/spec/support/../../ruby

# After
/appsignal-ruby/spec/diagnose/ruby
```