Test Push API key validation without color
The Elixir integration doesn't print the output with color (yet), so
update the regexp to also pass without color expressions.

Ideally we'd probably run these kinds of tests with a `--no-color`
option, and add additional tests to test the presence of color, but not
the value at the same time.