Update paths ownership regexp
The Elixir integration doesn't print the user/group name for the
ownership lines for the paths section.

I don't remember exactly why, but I think it's a lot more difficult in
Elixir to fetch the user/group name based on a user/group id. If we can
fix it in the Elixir integration we can revert this change, but I didn't
want to dive into that rabbit hole at the moment.