Update the config section for Elixir
The Elixir integration prints all its config values alphabetically,
that's how it stores Map objects internally. This means not even the
default Environment, debug and log config options are at the top.

The Elixir integration is also the only one that by default prints a
breakdown of the config option value sources, which we should add a test
for for all integrations.