Move runner prompt input to main Runner class
Instead of implementing the prompt handling per language runner,
implement it in the `Runner#run` method. Move the prompt to the Runner
initializer as a config option so we can customize it.

For some specs I don't want to use the prompt and I don't want it to
accidentally fall back on it. So the test that tests `--no-send-report`
does not have the prompt option set.