Move Runner arguments handling to language runners
Reverse of the prompt change in the previous commit. Add the arguments
of the diagnose tool to the language runners. This means it's
implemented multiple times, but this will be needed for the Elixir
integration.

The Elixir integration will need to use a release build of the app,
because I can't get it to not overwrite download and install report
every time the app is run. The release of the app has a different set of
argument handling so it cannot be put at the end of the command the same
way for every integration.