Update missing report file expectation for Elixir
The Elixir integration has two errors when the download and install
report files are missing. The error messages are also too different to
use the same matcher for it.

Ideally this would also test if only the download or install report is
missing, but the test for how the install report is rendered now doesn't
work yet at all.