Fix Elixir install report tests
The Elixir integration has two install reports: a download and install
report. The download report contains much of the same fields as the
install report. They are split up because the download and installation
can occur at different times.

This change updates the install report test to test for the additional
fields for Elixir.