Fix Elixir host architecture assertion
The Elixir host section returns the architecture as the full target
triple: `'x86_64-apple-darwin20.6.0'`.

Maybe we should update it to only return the architecture `x86_64`, but
there's no way to only get that from Elixir. We have a similar field in
the install report that's created in the `mix_helpers.exs`, but I'm not
sure we want to duplicate that logic at this time.