Test agent report in error scenario
When the Push API key is not defined the agent diagnose will report an
error. Test if the integrations all print the same error report.

To trigger this failure scenario I had to make the Push API key env
variable an optional setting. If it's set to an empty String, it will be
interpreted as not being set.