Submit diagnose reports to spec server
To test the diagnose reports for all integrations, submit the reports
from the integrations to a mock server for the test suite.

This mock server will receive the report and parse the JSON. Then we can
use that parsed report to test on it in the diagnose tests.

## Changes to the specs

I've added separate specs for report structure and values. I didn't want
to mix them in the output specs even though they test the same values,
but I don't want to spec to fail if only the output or the report is
wrong.

I've added a new `expect_report_for` helper to test the report itself,
but this is not done yet, as it doesn't make sure that the entire report
section structure matches.

The specs now submit the diagnose report by default, and we have
separate tests for what happens when we don't. This makes it easier to
test the report values along side the output specs.

## Changes to the integrations

To have the integrations submit to this mock server, they need to be
configured with a new server endpoint. I've done this with the
`APPSIGNAL_DIAGNOSE_ENDPOINT` environment variable. The integrations
will need to be updated to listen to this environment variable and use
its value as an endpoint instead of the default `appsignal.com/diag`.

## Output to STDOUT

The diagnose specs now log Sinatra requests (from the mock server) to
STDOUT. I've already set `logging` to `false`, but it still logs some
information. Not sure how to disable that.