Fetch the report from nested diagnose key
The diagnose CLIs in the Ruby and Elixir integrations wrap the diagnose
report in a "diagnose" key. This is because of how the Rails endpoint
was originally written.

Update the diagnose tests, which were originally written against the
Node.js integration, to fetch the report from the nested diagnose key.
I've updated the Node.js integration to match.