Update diagnose specs to match Ruby gem format
I consider the Ruby gem leading in most of the data that the report
includes, so I've updated it to expect the Ruby gem's format.

The other integrations will need to match it. The only exception I know
is the `user` and `group` keys for path ownership, it's not easy to get
these in all languages, so they're only required for the Ruby gem.