Fix Elixir install report stub
When we would release a new version of the Elixir package and then test
against it, it would create a directory in the build release directory
for that package version. Over time multiple versions could be in this
build release directory, causing the report stub behavior in
`after_setup` to be confused which package version to stub it in. It
would pick the first version, assuming there was only one directory.

I didn't want to do something like find the latest version of the
package, because that's not necessarily the one someone is currently
testing against.

Remove any AppSignal version package from the build release directory to
make sure it's always a clean state and it can't get confused about
which package version to write the report stubs to.