Update install report specs for Elixir
The install report for Elixir contains some fields other integrations do
not. Some of these should be added to the other integrations as well,
but because the Elixir integration has a download and install report
some fields will be duplicated. These duplicate fields are not necessary
for the other integrations.