Test each config sub-section independently
To separate the test for the `options` section of the config from
the one for the `sources` section of the config, the `section`
method in the `DiagnoseReport` is modified so that it takes a
series of keys to pass to `Hash#dig` on the report, instead of
a key to pass to `Hash#fetch`.

The implementation-pending assertions on the sources are moved to
a separate test.