Add config sourced from file
In order to test the behaviour when configuration variables are
set in the configuration files for each platform, configuration
files are added to the `ruby` and `elixir` project folders.

The `enable_minutely_probes` key is modified within these files as
well, in order to ensure overrides from the defaults are tested.