Expect params' `name` to be non-empty
In PR #30, a test that checks that the name is sent as empty in
the diagnose request params was added, while in PR #34, a config
file that makes the name not empty was added for the Ruby and
Elixir test projects.

This commit aligns the expectations in #30 with the changes in #34
by expecting the name to be set in Ruby and Elixir.