Ensure env is always development
This commit ensures that the `environment` configuration value
is always set to `development` for all integrations. For the
Node integration, it sets `NODE_ENV` instead of setting
`APPSIGNAL_APP_ENV`.

To do so, the way environment variables are handled in the
`Runner` and its subclasses is modified, with a new `run_env`
function that subclasses can implement. `run_command` is also
modified so that it takes the command-line arguments as an
argument.

In order to continue testing the behaviour for the Node diagnose
output when multiple configuration sources have values for
the same configuration key, and in concordance with the
behaviour of the other integrations, the `enable_minutely_probes`
key is set to `false`, overriding its default value.