Ruby move send_session_data to the override source
For the Ruby gem we now track `send_session_data` in the `override`
config source. The system source is too early in the config load order
and it was previously set much later, which meant it could result in an
inaccurate diagnose report.

Part of https://github.com/appsignal/appsignal-ruby/issues/811