Fix formatting for Python agent tests
It would print things like `False`, `not started`, and `not writable`
for tests that weren't present. This doesn't mean they are not writable,
only that we couldn't test it. Most likely because the config wasn't
loaded and it didn't know what paths to check.

Fix the format to match the other integrations and print a dash instead.