Assert more details about paths in Python
The Python diagnose CLI didn't print all the path details other
integrations do. This change now asserts that it does.

Not sure why on Node.js the "Log directory" doesn't output the same
details, but that's out of the scope of this change for Python.
Otherwise I would have removed that if-statement around the log
directory's assertions.