Add running_in_container to agent diagnose report
For the Python package we need a way to communicate if the app host is a
container or not. The agent diagnose report is used for this, update the
matchers to match.

This field will be present for all integrations, not just Python.

Part of https://github.com/appsignal/appsignal-python/issues/172