Fix Python diagnose working dir and initial config
In a recent version of hatch, they made sure the working directory stays
the same for `hatch run`. It will now be the directory in which the
`hatch run` was run, not be the project directory of the package you're
running.

Update the paths to match the working directory of the diagnose tests
submodule.

Now that it uses the correct working directory, it also loads the
`__appsignal__.py` config file correctly, so update the initial config
source content. Remove the ignored line that hid this issue.