Fix bundler env being copied to Ruby runner
Fix issues with the wrong bundle of gems being used as context for the
runner. This failed to find the Ruby gem as it was using this project's
test suite Gemfile instead.

Wrap commands with `Bundler.with_unbundled_env` so it removes any
Bundler context from the environment. It also clears some other env
vars, so explicitly set them again in the base `run_env` method.